Refactored combat module to be more DRY
* Was repeating too much code in some aspects, condensed down and
extracted more common elements.

* Preparing to allow complex chainings and combat phases

+++ OpenORPG.Server/Game/Entities/Character.cs
@@ -183,7 +183,7 @@ namespace Server.Game.Entities
         {
             var combatSystem = Zone.GetGameSystem<CombatSystem>();
             if (combatSystem != null)
-                combatSystem.ProcessCombatRequest(this, skillId, targetId);
+                combatSystem.ProcessCombatRequest(this, skillId, (int) targetId);
         }
